// src/app/shared/infrastructure/api/generated/model/walletRechargeResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Backend-WebApplication
 * ChapaTuRuta REST API â Identity & Access Management, Driver Management, and Monetization
 * OpenAPI spec version: 1.0.0
 */
import type { WalletResponse } from './walletResponse';
import type { WalletTransactionResponse } from './walletTransactionResponse';

export interface WalletRechargeResponse {
  wallet?: WalletResponse;
  transaction?: WalletTransactionResponse;
}