// src/app/shared/infrastructure/api/generated/model/walletTransactionResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Backend-WebApplication
 * ChapaTuRuta REST API Ã¢ÂÂ Identity & Access Management, Driver Management, and Monetization
 * OpenAPI spec version: 1.0.0
 */

export interface WalletTransactionResponse {
  id?: string;
  walletId?: string;
  tripId?: string;
  type?: string;
  amount?: number;
  resultingBalance?: number;
  timestamp?: string;
}