// src/app/shared/infrastructure/api/generated/model/farePolicyResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Backend-WebApplication
 * ChapaTuRuta REST API â Identity & Access Management, Driver Management, and Monetization
 * OpenAPI spec version: 1.0.0
 */

export interface FarePolicyResponse {
  id?: string;
  baseFare?: number;
  pricePerKm?: number;
  minimumFare?: number;
  commissionRate?: number;
  updatedAt?: string;
}